+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added event handling on dashTile
 
 - Added Constructor-schemes component
+- Added loading svg
 
 ### Fixed
 - Fixed error messages fo password un profile
@@ -25,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add password max length
 - Added visualization  circular dial
-- Added loading svg
 
 ### Fixed
 - Change username on edit