+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added event handling on dashTile
 - Upgrade yFiles to 24.0.5
 - Added Constructor-schemes component
+- Added loading svg
 
 ### Fixed
 - Fixed error messages fo password un profile